move collective communication function to dev_scripts. some cleanup
